% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Primary biliary cirrhosis data}
\format{
A data frame with 1945 robservations and 5 variables:
\describe{
\item{id}{Patient ID.}
\item{years}{number of years between registration and the earlier of death, transplantation, or study analysis time.}
\item{status}{Status with three levels \code{alive}, \code{transplanted} and \code{dead}.}
\item{year}{number of years between enrollment and this visit.}
\item{hepatomegaly}{A binary outcome used to indicate whether hepatomegaly is present or not.}
}
}
\usage{
pbc
}
\description{
A dataset containing the primary biliary cirrhosis data collected from January 1974 to May 1984. This dataset is a simplified dataset of \code{pbc2} in \code{JM} package.
}
\references{
\cite{Thomas R. Fleming, David P. Harrington (1991). "Counting Processes and Survival Analysis." Wiley, New York.}

\cite{Peter Hall, Hans-Georg Müller, and Fang Yao (2008). "Modelling Sparse Generalized Longitudinal Observations with Latent Gaussian Processes." Journal of The Royal Statistical Society Series B-statistical Methodology, 70(4):703-723.}
}
\keyword{datasets}
